Updates to address user comments
Updated create-flow.md article to remove text about "new" or "updated" features.

Added information from create-flow.md to point to the article about configuring flows, because that is the article where the customer question was already answered.

Did some cleanup in cohort-analysis.md, as well as t-cohort.md. There was some funky formatting displaying in the published article.

Also added section on downloading the cohort visualization that points to the correct information.