Updates to sharing documentation
Added information about how to view shared projects when they are shared with you.
Streamlined information about sharing individual or multiple projects. (There was a separate section in this article that explained how to share multiple projects, but it was a bit convoluted. I combined the information into the section about sharing a single project.)