Removed derived metric terminology
Removed all description of derived metrics from AA docs. Confirmed with Derek Tangren that derived metric is only used as a concept in CJA as part of configuring a metric in a data view. There is no such concept in the context of calculated metrics, both in AA and CJA.